# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API untuk platform validasi hasil kerja AI oleh ahli manusia.",
        "title": "AIValid API",
        "termsOfService": "https://aivalid.id/rules-content",
        "contact": {
            "name": "AIValid Support",
            "url": "https://aivalid.id/contact-support"
        },
        "license": {
            "name": "Proprietary"
        },
        "version": "1.0"
    },
    "host": "api.aivalid.id",
    "basePath": "/api/v1",
    "paths": {
...
        "/account/telegram/connect": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Link a Telegram account to the authenticated user using Telegram Login Widget data. The payload is verified against the bot token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "Connect Telegram account",
                "parameters": [
                    {
                        "description": "Telegram login widget payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerTelegramConnectRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerTelegramConnectResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Telegram account already linked to another user",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/account/telegram/disconnect": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Unlink the connected Telegram account from the authenticated user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "Disconnect Telegram account",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerTelegramConnectResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/admin/users/{userId}/badges": {
            "post": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "Assign a badge to a user by user ID. Admin authentication required.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-Users"
                ],
                "summary": "Assign badge to user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Badge assignment data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.AssignBadgeRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerUserBadgeAssignResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User or badge not found",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Badge already assigned",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/users/{userId}/badges/{badgeId}": {
            "delete": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "Revoke (remove) a badge from a user. Admin authentication required.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-Users"
                ],
                "summary": "Revoke badge from user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Badge ID",
                        "name": "badgeId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerMessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User badge not found",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.AssignBadgeRequest": {
            "type": "object",
            "required": [
                "badge_id"
            ],
            "properties": {
                "badge_id": {
                    "type": "integer"
                },
                "reason": {
                    "type": "string"
                }
            }
        },
...
        "handlers.SwaggerErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "AUTH001"
                },
                "details": {
                    "type": "string",
                    "example": "Detail tambahan"
                },
                "message": {
                    "type": "string",
                    "example": "Email atau password salah"
                }
            }
        },
...
        "handlers.SwaggerMessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Operasi berhasil"
                }
            }
        },
...
        "handlers.SwaggerTelegramAuthInfo": {
            "type": "object",
            "properties": {
                "connected": {
                    "type": "boolean",
                    "example": true
                },
                "deep_link": {
                    "type": "string",
                    "example": "https://t.me/johndoe"
                },
                "display_username": {
                    "type": "string",
                    "example": "@johndoe"
                },
                "first_name": {
                    "type": "string",
                    "example": "John"
                },
                "last_name": {
                    "type": "string",
                    "example": "Doe"
                },
                "photo_url": {
                    "type": "string",
                    "example": "https://t.me/i/userpic/photo.jpg"
                },
                "telegram_user_id": {
                    "type": "string",
                    "example": "123456789"
                },
                "username": {
                    "type": "string",
                    "example": "johndoe"
                },
                "verified_at": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                }
            }
        },
        "handlers.SwaggerTelegramConnectRequest": {
            "type": "object",
            "properties": {
                "auth_date": {
                    "type": "integer",
                    "example": 1700000000
                },
                "first_name": {
                    "type": "string",
                    "example": "John"
                },
                "hash": {
                    "type": "string",
                    "example": "abc123def456"
                },
                "id": {
                    "type": "integer",
                    "example": 123456789
                },
                "last_name": {
                    "type": "string",
                    "example": "Doe"
                },
                "photo_url": {
                    "type": "string",
                    "example": "https://t.me/i/userpic/photo.jpg"
                },
                "username": {
                    "type": "string",
                    "example": "johndoe"
                }
            }
        },
        "handlers.SwaggerTelegramConnectResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "ok"
                },
                "telegram_auth": {
                    "$ref": "#/definitions/handlers.SwaggerTelegramAuthInfo"
                }
            }
        },
...
        "handlers.SwaggerUserBadgeAssignObj": {
            "type": "object",
            "properties": {
                "badge_id": {
                    "type": "integer",
                    "example": 3
                },
                "granted_at": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                },
                "granted_by": {
                    "type": "integer",
                    "example": 1
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "reason": {
                    "type": "string",
                    "example": "Excellent contributor"
                },
                "user_id": {
                    "type": "integer",
                    "example": 42
                }
            }
        },
        "handlers.SwaggerUserBadgeAssignResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Badge berhasil diberikan"
                },
                "user_badge": {
                    "$ref": "#/definitions/handlers.SwaggerUserBadgeAssignObj"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AdminAuth": {
            "description": "Admin JWT token. Type \"Bearer\" followed by a space and admin JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}